Restore client hydration for filename-matched modules
The previous fix normalized MDX ESM cache entries, which fixed server import failures, but browser hydration still fetched project modules directly through /_vf_modules. That route transforms source on demand and could return only a named PlatformOverview export, causing the browser default import to fail before footer interactions could hydrate.

This shares the filename-based default export synthesizer across both paths and applies it to module-server output after transformToESM. Browser module tests now cover direct named exports and barrel exports for /_vf_modules/components/PlatformOverview.js.

Constraint: Studio preview imports filename-matched project modules as default imports during hydration

Rejected: Patch only the Tomcode project | preview runtime must handle this export shape for any project

Confidence: high

Scope-risk: moderate

Tested: deno test --no-check --allow-all src/modules/server/module-server.test.ts src/transforms/mdx/esm-module-loader/module-fetcher/module-cache.test.ts

Tested: deno check src/modules/server/module-server.test.ts src/modules/server/module-server.ts src/modules/loader-shared/filename-default-export.ts src/transforms/mdx/esm-module-loader/module-fetcher/module-cache.ts src/transforms/mdx/esm-module-loader/module-fetcher/module-cache.test.ts

Tested: deno lint src/modules/server/module-server.test.ts src/modules/server/module-server.ts src/modules/loader-shared/filename-default-export.ts src/transforms/mdx/esm-module-loader/module-fetcher/module-cache.ts src/transforms/mdx/esm-module-loader/module-fetcher/module-cache.test.ts deno.json src/utils/version-constant.ts

Tested: git diff --check
