+++ src/transforms/mdx/esm-module-loader/module-fetcher/module-cache.ts
@@ -17,78 +17,7 @@ import { hashString } from "../utils/hash.ts";
 import { buildMdxEsmModuleFileName, buildMdxEsmPathCacheKey } from "../cache-format.ts";
 import { hasUnresolvedImports } from "./nested-imports.ts";
 import { recordModuleToSession } from "./render-sessions.ts";
-
-const IDENTIFIER_PATTERN = /^[A-Za-z_$][\w$]*$/;
-const DEFAULT_EXPORT_PATTERN = /\bexport\s+default\b|\bexport\s*\{[^}]*\bas\s+default\b[^}]*\}/;
-
-interface ExportMatch {
-  localName: string;
-  sourceClause: string;
-}
-
-function escapeRegex(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function getModuleIdentifier(normalizedPath: string): string | null {
-  const fileName = normalizedPath.replace(/\?.*$/, "").split("/").pop();
-  const identifier = fileName?.replace(/\.(?:[cm]?[jt]sx?|mdx)$/i, "");
-  if (!identifier || !IDENTIFIER_PATTERN.test(identifier)) return null;
-  return identifier;
-}
-
-function findExportMatch(moduleCode: string, exportName: string): ExportMatch | null {
-  const declarationPattern = new RegExp(
-    `\\bexport\\s+(?:const|let|var|function|class)\\s+${escapeRegex(exportName)}\\b`,
-  );
-  if (declarationPattern.test(moduleCode)) return { localName: exportName, sourceClause: "" };
-
-  const exportListPattern = /\bexport\s*\{([^}]*)\}\s*(from\s*["'][^"']+["'])?\s*(?:;|$)/gm;
-  let match: RegExpExecArray | null;
-  while ((match = exportListPattern.exec(moduleCode)) !== null) {
-    const specifiers = match[1]?.split(",") ?? [];
-    const sourceClause = match[2] ? ` ${match[2]}` : "";
-    for (const rawSpecifier of specifiers) {
-      const specifier = rawSpecifier.trim();
-      if (specifier === exportName) return { localName: exportName, sourceClause };
-
-      const alias = specifier.match(/^([A-Za-z_$][\w$]*)\s+as\s+([A-Za-z_$][\w$]*)$/);
-      if (alias?.[2] === exportName && alias[1]) {
-        return { localName: alias[1], sourceClause };
-      }
-    }
-  }
-
-  return null;
-}
-
-function appendExportBeforeSourceMap(moduleCode: string, exportStatement: string): string {
-  const sourceMapMatch = moduleCode.match(/\n\/\/# sourceMappingURL=.*$/);
-  if (sourceMapMatch?.index === undefined) {
-    return `${moduleCode.trimEnd()}\n${exportStatement}\n`;
-  }
-
-  const beforeSourceMap = moduleCode.slice(0, sourceMapMatch.index).trimEnd();
-  const sourceMap = moduleCode.slice(sourceMapMatch.index);
-  return `${beforeSourceMap}\n${exportStatement}${
-    sourceMap.endsWith("\n") ? sourceMap : `${sourceMap}\n`
-  }`;
-}
-
-export function ensureFilenameDefaultExport(normalizedPath: string, moduleCode: string): string {
-  if (DEFAULT_EXPORT_PATTERN.test(moduleCode)) return moduleCode;
-
-  const exportName = getModuleIdentifier(normalizedPath);
-  if (!exportName) return moduleCode;
-
-  const exportMatch = findExportMatch(moduleCode, exportName);
-  if (!exportMatch) return moduleCode;
-
-  return appendExportBeforeSourceMap(
-    moduleCode,
-    `export { ${exportMatch.localName} as default }${exportMatch.sourceClause};`,
-  );
-}
+import { ensureFilenameDefaultExport } from "#veryfront/modules/loader-shared/filename-default-export.ts";
 
 /**
  * Normalize a module path, resolving relative paths if a parent is provided.
